Fixed: Gravitation for forces for the beam element.
Calculate these by multiplying the acceleration of gravity
with the element mass matrix, such that we also get the
contributions to the rotational DOFs due to the Hermitian
interpolation functions on which the explicit element matrices
are based. They will normally cancel with the neighbor element
except for at the beam end where a bending torque may occur.